Tweak the wording and order in discussion of #51
Summary:
Our error docs are intended mainly as a reference for people who
actually encounter errors, and one reason I've been giving them a
once-over is because we're hoping to improve error message quality
and I think that good error docs - and maybe linking to them the
way flake8 does - could be handy.

So in this commit I change to lead with discussion of the cause of the error - mutliple modes - and then follow up with a context paragraph about what local modes are (including the links to full discussion elsewhere in the docs).

Reviewed By: dkgi

Differential Revision: D30091991

fbshipit-source-id: 3df90a1e6905fd5fe44be3ddd3f1fd17282ec849