Change the wording on unsafe cast
Summary:
Our error docs are intended mainly as a reference for people who
actually encounter errors, and one reason I've been giving them a
once-over is because we're hoping to improve error message quality
and I think that good error docs - and maybe linking to them the
way flake8 does - could be handy.

Based on this philosophy about how the docs are used, I think we
should lead with a discussion of the actual error - in this case
`UnsafeCast`, which is always produced from an unsafe call
to `safe_cast`. The discussion of `typing.cast` which is unsound
should be in a follow-up paragraph as context, so that readers aren't
left wondering what the point is for a full paragraph.

Reviewed By: dkgi

Differential Revision: D30091722

fbshipit-source-id: 9cafee45f3e6276c98b8617dd1487fb2c10248af