Rename Callable -> Target [1/n] rename module
Summary:
We decided recently to rename many of the interprocedural
modules to better reflect what they actually do, and also
minimize collisions with `Analysis` module names.

The one rename that we decided made sense but we put off at first was `Callable
-> Target`... this rename is desirable because (a) we already have non-callable
targets in taint and (b) there's an open wish-list item to allow tracking taint
through attributes.  But it will take quite a while to rename all the various
functions and variables that have `callable` in the name.

This is the first commit of a rename, and by far the biggest bulk commit we
need to make, it renames the module without changing anything else. After we
land this, I can go through this change line-by-line as well as use code search
to identify variables and functions that need renames in much smaller follow-up
diffs.

Reviewed By: grievejia

Differential Revision: D30184905

fbshipit-source-id: f5cedb9da53fa3e0280a1c857f339f6c7271fd1d