Rename hardcoded decorator functions to be non-private
Summary:
Verified that these names weren't meant to be private, just added a prefix to avoid conflicts.

Because inlined decorators generate new functions, this doesn't go through preprocessing a second time to make sure the newly generated (hard coded) function names get mangled properly if they are private.

When testing, the "expected" source is parsed with all expected generated functions already there and so these are mangled when they're named in a way that looks private.

https://docs.python.org/3/reference/expressions.html?highlight=mangling#atom-identifiers

> Private name mangling: When an identifier that textually occurs in a class definition begins with two or more underscore characters and does not end in two or more underscores, it is considered a private name of that class. Private names are transformed to a longer form before code is generated for them. The transformation inserts the class name, with leading underscores removed and a single underscore inserted, in front of the name. For example, the identifier spam occurring in a class named Ham will be transformed to _Hamspam. This transformation is independent of the syntactical context in which the identifier is used.

Could either throw in an exta re-preprocessing of the "actual" source after it's gone through decorator inlining, or just change the way we hardcode the generated names to avoid the name mangling (they're not even methods on the class, so it makes no practical difference). Figured just avoiding the mangling situation is cleaner; unblocks D30146161.

Reviewed By: dkgi

Differential Revision: D30212576

fbshipit-source-id: 2c4b47d91f3e478297fe68ae55413f8bbaced6b1