Print toplevel unbound names if sources unequal but statements identical
Summary:
A huge pain point for me when trying to debug a diff was running into tests that compared sources, where the two sources were ostensibly identical based on the test output and the printed diff between expected and actual was empty.

It turned out that sometimes the top level unbound names were different in the source, sometimes there was some statement node nested deep into the source that had some metadata that was different, etc.

It would have saved me a lot of time and print debugging if the source equality testing function could tell when it was not going to print any helpful output around why this test was failing, and adjust to print something useful instead.

Decided to implement this and land it - for now, if the sources are not equal even without counting the statements in the body, print the pretty json and compare that. Otherwise, just do the normal printing. Can be modified to maybe do pretty printing more aggressively if trying to debug in the future and something still seems to show no difference.

Reviewed By: grievejia

Differential Revision: D30213681

fbshipit-source-id: 6a9447f0301858c121e8a2ee98a7d3cc15144586