Remove special cased private attribute error logic
Summary:
This is now a no-op!

Attempts to access private attributes from outside the base class will result in not being able to find the attribute, which is already caught in general logic.

This match case was specifically for attributes that were successfully found and were defined, but we needed a layer of checks on the name and the parent to approximate a private attribute error.

For context: prior to the mangling change, this case was necessary to throw errors on attribute access where the attribute *was* found defined. There is actually separate error throwing logic for the target of assignments in forward statement that did not have this special casing, which lead to attribute setting of private variables to throw no error while accesses did.

Reviewed By: dkgi

Differential Revision: D30241080

fbshipit-source-id: ac9239eb257078ee8820806a089d9d02c31a23fe