Refactor/decouple: coverage collector
Summary:
Move coverage collector into a file of its own, and decouple it from statistics.

From code-flow/code sharing point of view, it is already not coupled with statistics, so this is a straightforward change. At an earlier point, I was interested in it sharing more logic with statistics that is why had put the `CoverageCollector` in the `statistics_collector` module. But have now decided to keep coverage collection independent, so it wouldn't raise concerns of affecting the statistics collection.

With that decision, it makes sense to put this in a file of its own.

Reviewed By: shannonzhu

Differential Revision: D30168548

fbshipit-source-id: 7fcb0eed623d9c1dc578e0647a0aca32d19c9d9d