Rename raw ast class bases as base_arguments
Summary:
To avoid confusion in the future, I'd like to make it very clear that things that syntactically go in the "arguments" portion of a class definition are *not* just the bases of the class.

This made more sense when we assumed everything in there had to be a base class or a metaclass, which is a similar concept, but now they can be arbitrary arguments into a parent `__init_subclass__` method.

Calling these bases is no longer accurate; they're just arguments to the class in the AST, some of which can be parsed as bases.

Reviewed By: grievejia

Differential Revision: D30263256

fbshipit-source-id: 3b0c0e27e08efb994f04c7a4bdcf7dd4155b8183