Be explicit in errors if attribute is private and undefined
Summary:
If we can't find a private attribute and it follows the naming pattern that would get mangled, let's add some context to the error message to flag to the user that privacy is one likely reason the attribute they're trying to access is undefined.

Note: string checks are the best we can do here, unless we actually want to re-resolve for an attribute of a matching mangled name up the inheritance tree and only add this if such an attribute exists. Figured that's probably not worth the perf hit.

Reviewed By: grievejia

Differential Revision: D30253573

fbshipit-source-id: 4740b628802c53d520045df9b7ffcb8085f4a59e