Add --no-use-command-v2 flag
Summary:
If you set a click flag with a default that is not False
and no off-switch, it is not possible to set False. As a
result, the `--use-command-v2` flag (which defaults to
`None` meaning "use what the conf file says) cannot be
explicitly disabled.

I'd like an explicit off switch for use cases where CI
calls into commands that are newly-adapted to the new
server so that we can easily set a feature flag to roll
back if there are problems.

Reviewed By: grievejia

Differential Revision: D30289237

fbshipit-source-id: 96a1501ddd4ebb42b4de98da01ab805b096ac387