Pyre coverage: make paths relative to project root
Summary:
Make the paths in `pyre coverage` output relative. This makes it more concise and reasonable looking.

More importantly, it makes the output more portable (can be interpreted without post-processing in a new checkout/context where the full path of the project root would be invariably different).

There are a few different options, what this could be relative to:
* relative to project root
* relative to local root
* relative to current working directory

We would like to be able to combine outputs of coverage from different local roots, that is why we are not going that direction (it can be made to work, but will require more post-processing to combine).

Current working directory could also be made to work, but I don't find it particularly appealing at this point.

Reviewed By: shannonzhu

Differential Revision: D30296152

fbshipit-source-id: 15922e4828867504d48a3c5d1dea59e665e0f225