Get lines spanned by format string.
Summary:
Problem: When suppressing errors during upgrades, we need to insert the fixme comment at the start of a multi-line format string. Right now, we try to insert within the format string, which causes an AST change, and makes us abort the entire upgrade.

This diff gets the lines spanned by format strings. The next diff will use this data.

Reviewed By: shannonzhu

Differential Revision: D30296837

fbshipit-source-id: 8b29ee2eee09d27a483322b217717fce519f1e86