Collect format strings that have a fixme at the beginning.
Summary:
Problem: f-strings often span multiple lines and have some error in them. However, we can't suppress those errors because it is within an f-string and inserting a fixme there would change the behavior (and the AST). Our upgrade script rejects such AST changes. This means many projects couldn't be upgraded.

Solution: We will spread the fixme at the start of the f-string to apply to all the lines spanned by the f-string.

Pros:
+ Pretty user-friendly. The mental model remains: put the fixme comment at the start of the expression. No need to learn anything new for multiline f-strings.
+ Will allow us to suppress errors within f-strings, something we cannot do right now.

Cons:
+ This approach will suppress any error of the same code within the entire f-string. The alternative is to somehow specify the line number in the fixme comment, which doesn't seem worth it.
+ We might suppress an error on the last line if there is some other expression (say, with a semicolon). Will go into this in the diff where I make the change.

The pros outweigh the minor cons, imo.

Reviewed By: shannonzhu

Differential Revision: D30294164

fbshipit-source-id: 7c50569a92437c071e5ec1903305c6be9c3aa8e7