Test case for instantiation of attributes via init subclass
Summary:
This now works out of the box. Adding testing for this specific user concern around our previously missing support for init subclass.

Note: we are actually overcorrecting a bit on the errors here, as we haven't added special casing that ensures attributes instantiated in `__init_subclass__` do not count for the current class - only for children. I've filed T98000466 to track this, but won't spend more time on it now.

Reviewed By: dkgi

Differential Revision: D30270737

fbshipit-source-id: 4d124ae228680579842b150cbc02e9498a172f3b