Add to whitelisted dunder methods from override errors
Summary:
Not positive what makes the cut to be whitelisted here, or if there is a consistent rule for the ones that already exist in this list. See the diff that introduced this: D21583772 (https://github.com/facebook/pyre-check/commit/65da23e25db99591a3dd2883c78d374464a372ea)

If a class overrides a standard magic method (ie., `__iter__`, `__delitem__`, `__lt__`, `__add__`) and is properly consistent with the parent function in every way but the parameter naming, should we be more lenient? Especially because the naming convention used in the `object` class for some of these is not very descriptive, ie., parameters named `o`, `v`, `n`, `i`, etc.

RFC: Since by design these are almost never called by an explicit method invocation, should we avoid forcing users who define these methods to use the same parameter names?

Some discussion:

https://www.internalfb.com/diff/D29890102?transaction_fbid=3717297901703058

> It is technically true that it's inconsistent, because I'm using key and the base class is using v as the parameter name. But it's easy to argue that this inconsistency has no practical significance, because the only way I could run into an issue is if I were to use it as a keyword argument, e.g.:

`d = dict() if use_dict else frozendict()d.__delitem__(v="foo")`

> Arguably, nobody will do that. People will just write del d["foo"].

Note: This diff adds a couple magic methods to a whitelist that ignores all override-related errors. Alternatively, we could move the whitelist check to only gate `StrengthenedPrecondition` of `NotFound` variant to only avoid erroring on misnamed parameters, but doing so results in a lot of errors where we also don't want to error on inconsistent overrides of the return parameter on `__call__` for example.

Reviewed By: dkgi

Differential Revision: D30308116

fbshipit-source-id: 2b4566c169c8869a84b769a99220bec1915537dc