Don't error on incomplete overload if implementation exists
Summary:
Noticed some strange spurious errors on overloads not being general enough to cover the signature they were overriding.

Most of the time, overloaded functions will break down a more general signature into components, each function overload accepting a subset of the acceptable parameter types and returning a subset of the possible return types.

Currently, Pyre is erroring if a subclass identically defines a method and its overrides - because it sees one of the overrides and notices that it's not broad enough to cover the implementation in the parent. (for example, see the very first unit test added in this diff. we used to error here even though the subclass defines each method exactly the same way as in the parent).

This diff proposes holding off on incompatible override errors if we're looking at an overload, instead only erroring if (1) the current define is an implementation, and doesn't fully cover the overridden implementation, or (2) we see a bunch of overload functions with no following implementation.

This should cover most unsafe cases without a lot of false positives on overload overrides.

Reviewed By: kbansal

Differential Revision: D30343187

fbshipit-source-id: e834db31e36d9a55d872d53fb3131dedc82138ac