Use full annotation store in resolution fixpoint
Summary:
Store and fetch the full set of local annotations, as we used to.

iirc, the resolution fixpoint is used so we can rebuild a resolution and have typing context at each step in the fixpoint / before and after each statement.

Before this diff to make the temporary annotations change a little easier to tackle at once, this continued to only store non-temporary annotations.

Temporary annotations *are* part of the type context though, so we should both store the full annotation store (both permanent and temporary local annotations) when we save a pre/post condition, as well as read in both when constructing a resolution from a saved pre/post condition set.

Reviewed By: grievejia

Differential Revision: D30413804

fbshipit-source-id: 59139cfc1fa855a70a6654da42d119666af9095d