Error on re-annotating global identifiers
Summary:
This diff adds illegal annotation target errors when trying to re-annotate an already annotated global.

On master, we already error for a re-annotation of a local:

```
def foo() -> None:
    x: int = 1
    x: str  # error is on this line
    x = ""
    return
```
Illegal annotation target [35]: Target `x` cannot be annotated after it is first declared.

But we don't do this for re-annotation of a global:

```
constant: int = 1

def foo() -> None:
    global constant
    constant: str  # no error
```
And instead relied on the fact that we were special casing out any assignments that touch globals from changing the local annotation map.

In this case, we should be throwing the same error on re-annotation of an already declared global, which would be a lot more user friendly than silently letting the re-annotation pass without actually changing the type, then only erroring if we try to assign a value later.

Reviewed By: grievejia

Differential Revision: D30404490

fbshipit-source-id: 21cb9beb60bb651cb5fa36c6e96a964bfdfa80ed