Use local mode in infer unit tests
Summary:
I expect to fully delete infer interprocedural mode, as it
currently exists using the interprocedural framework built for
taint analysis, in the next few weeks - it's dead and the code
sitting there is just complicating things.

We may eventually still use the infer mode to turn on a form of
interprocudural inference based on AST rewriting and incremental-style
re-analysis, but that's probably not in the near future, and in the
meantime I don't think we should be using non-local modes in unit
tests.

Reviewed By: shannonzhu

Differential Revision: D30352061

fbshipit-source-id: 3f65d0c8a0093669adab297510ef6f32c191cc8d