Remove invalid refinement todos
Summary:
On second glance these behaviors are actually fine.

In the first test, the interleaving call only happens if the attribute is None, so if we successfully refined the attribute that call won't happen and the temporary refinement shouldn't be cleared.

In the second test, short circuiting will prevent the interleaving call from actually happening in the if check. (and checks won't refine anything at all, since no guarantee any specific piece of it held - as seen in preceding test).

There are some separate cases where we should be lenient, as documented here: T99107540

Reviewed By: dkgi

Differential Revision: D30555413

fbshipit-source-id: 70de5f237729419ab41c8d1da106d51b123e87b2