Infer: pass paths to annotate [1/3] python-side
Summary:
A growing problem with infer codemod commands is that they time out. The
root cause of this is also a major pain point when debugging: whenever we
try to annotate a single module, we run type inference on the entire project
which is extremely wasteful for a large project.

Now that we can use the new server for the ocaml side of infer, it should
be possible to pass `paths_to_modify` down from the client, enabling us
to just do the work required. This will probably make deugging an order of
magnitude faster on many projects, as well as fixing the timeouts in sandcastle.

This is the first part - adding the new field to the python-side arguments. I
want to do this first so I can prove that it's safe to land without worrying about
the binary release (this should be true because the ocaml deserialization ignores
unknown fields; having a new one should be fine).

Reviewed By: shannonzhu

Differential Revision: D30344269

fbshipit-source-id: ec5047ca987c8415d688124d64f60448ff76f446