Infer: pass paths to annotate [0/3] force absolute paths
Summary:
This is some pre-work for a push I'm making to push path filtering
down into the ocaml client, which will make it run much faster.

The prework is to maximize determinism and minimize the amout of stateful logic
happening in the `infer.py` module, by determining the working directory in the
outermost click handler and using absolute paths everywhere else.

Since the ocaml side only deals with absolute paths, this will also make it
much easier to push the filtering logic down.

Reviewed By: shannonzhu

Differential Revision: D30351791

fbshipit-source-id: 07d06da543d6fe83de799d54186d57cfbe7962f5