Fix potential loss of models when output generated models to stdout
Summary:
Due the the fact that we exclude annotations from hashing/equality for models (D20408377 (https://github.com/facebook/pyre-check/commit/eedcca49a7415e15c867e5df26db7eda8dadf9af)), putting two models for the same class/function/attribute into a set will cause one to be discarded. This can be a problem if two model generators produce different annotations for the same target, and we then combine the output of those model generators into a set. This was being done in cases where an explicit `--output-directory` flag was not provided.

This diff fixes that issue by not combining the sets. It also adds tests for this new behaviour, since this code was previously untested.

Reviewed By: arthaud

Differential Revision: D30543667

fbshipit-source-id: 3703d01dc111f519555a7573e21513af689772e4