Reafactor typing imports logic
Summary:
Prior to this commit, we had a general `split` method on annotations,
which was exclusively used to split possibly complex types into
simple types in order to then dig around for typing imports.

This commit hides the general `split` logic in a protected and more
clearly-named method `_simple_types`, and only exposes the `typing_imports`
logic, in the process ridding us of a top-level function that isn't really
needed.

**Context**

I want to add support to force pyre into using string annotations; this
is needed when preprocessing code for pysa because otherwise the
imports we create cause line counts in sapp to be off relative to checked-in
code.

It wasn't obvious what the implementation of "split" ought to be in that
case, but once we focus on the only actual use case, `typing_imports`, it
becomes quite obvious: in string annotation mode, there are never typing
imports.

I'd actually like to get rid of typing imports altogether, but there's a
bug in our libcst code generator that I'll have to address first (it
doesn't know how to infer imports from generic type names), and getting
a libcst release out is slow so we'll have to work around it for a while.

Reviewed By: kbansal

Differential Revision: D30581232

fbshipit-source-id: 0dbbb9dafa17467d900e484386dd0bc189214e3c