Give pyre infer the ability to quote annotations
Summary:
This commit adds a --quoted flag to pyre infer (only in the
v2 version of the command; I'm no longer making changes to the
old infer except for critical bugfixes).

**Why?**

When using pyre to preprocess code for pysa analysis (as opposed
to using it to make codemods we would want to check in), we
should place utmost priority on keeping the pysa results usable.

As a result, we want to avoid ever changing line numbers, since
they cause the sapp code to point at incorrect lines of committed
code in issues.

A few not-immediately-obvious considerations handled here:
- typing imports should always be empty when using quoted
  annotations - a quoted annotation never generates an import
- both the dequalify and use_future_annotation options are
  incompatible with the quote option in that no user would
  realistically want to mix them. We validate this in __post_init__
  to avoid silently producing weird results.

Reviewed By: pradeep90

Differential Revision: D30583958

fbshipit-source-id: 23c49eec7d19a24dcb0a2818e3984316d00d97fa