Raise error when reaching the fixpoint widening threshold for complex functions.
Summary:
Problem: We silently set variable types to be `unknown` when a function is very "complex". This specifically happens when we reach the fixpoint widening threshold. This can hide some serious bugs and confuse users.

Design choice: I considered returning a flag from `RefinementUnit.widen` to indicate whether the threshold was reached. But we use that function indirectly within `Map.merge`, which means I'd have to thread the information through all the calls using a mutable variable. Seemed like unnecessary work. So, I just checked for the threshold directly.

Reviewed By: dkgi

Differential Revision: D30561510

fbshipit-source-id: 5af486812c93346143bfa2e72b933c77cb86c94c