Remove the pre-new-server `pyre infer`
Summary:
The `pyre infer` command is under very active development; it
has a large number of open bugs, most of which are on the python
and libcst-integration side.

As a result, it's not practical to maintain two versions of the
infer python-side logic right now. Moreover, `pyre infer` is not
at all part of the typical commandline / CI / VSCode workflow
for normal code editing, it's really only used for pysa preprocessing
and codemods, which means any state inconsistencies between
new-server and legacy aren't relevant.

So, let's just get rid of the legacy code to avoid possible issues
caused by the legacy logic not being maintained.

Reviewed By: grievejia

Differential Revision: D30639730

fbshipit-source-id: 13b1bc64dfed9d84492659b7c1a29bb28be12d34