Treat code as unreachable after `assert <false> and x`
Summary:
Problem: We were treating code as unreachable after `assert <false>`. However, when it came to `assert <false> and x`, we were not doing so.

This was because we were falling back to the original resolution if `forward_statement <condition>` returned resolution as `None`. That meant that we were going into unreachable code and producing confusing results.

Solution: Checking `assert <false> and x` should return `None` as the resolution.

Note: I looked into the `Or` branch but could not come up with a test that would make the current code fail. So, I'm leaving it as is for now.

Reviewed By: shannonzhu

Differential Revision: D30650520

fbshipit-source-id: b21a1bd812b8a1c857cf57e7ee1db8f840003488