Line-level coverage information match file-level statistics (1/3)
Summary:
### Context

We already surface file-level coverage information, this is collected by  AnnotationCountCollector in statistics_collector. We are now interested in providing finer-grained coverage information that is line-level, this is collected by CoverageCollector in coverage_collector.

The existing implementation of CoverageCollector was a placeholder one with a trivial implementation. This stack of diff is intended to bring the line-level coverage information match the file-level statistics that we collect.

### This diff
As a first step towards it, collect information in CoverageCollector in a fashion similar to AnnotationCountCollector, but more granularly. So instead of counting, we are actually collecting the items that we were counting in AnnotationCountCollection. For testing, we verify that the lengths of the items we are collecting are the same as would have been with AnnotationCountCollector.

### Stack-to-be

The diffs in the stack that will follow will do the following:
1. Change the `covered_lines` and `uncovered_lines` be based on globals, attributes, parameters, etc.
2. Refactor CoverageCollector and AnnotationCountCollector, so that we have base class AnnotationCollector and share more code. We intentionally do not do this first, so as not affect statistics jobs.

Reviewed By: grievejia

Differential Revision: D30622246

fbshipit-source-id: 05479efe0eb2867959cb7d39e6f6606127abe9c4