Add syntactical support for `type_annotation` clause for attribute models
Summary:
As titled, this updates the model parser to actually parse clauses of this type into `AnnotationConstraint`s (added in D30550464). Note that the `type_annotation` clause can only be used when the find clause type is `"attributes"`, or within the parameter where clause, it is not supported for callables. (see `any_parameter` and `return_annotation` for use of annotation constraints with callables).

Added unit tests and integration tests for the new functionality.

Reviewed By: arthaud

Differential Revision: D30550462

fbshipit-source-id: 2e2986275cd7e00f82e0f1c029d7e30ddf0a79aa