Fix pyre infer python / ocaml bridge, improve type safety
Summary:
This change decouples the `in_place` flag from the
nullarity of `modules_to_annotate` - it was never a great idea to
use the `None`-ness of an optional set to indicate in place or not,
and this led to us shipping an incorrect change that makes infer a no-op when
we don't provide any paths to annotate.

**Context**

The changes in D30351792 (https://github.com/facebook/pyre-check/commit/05d9816075c42a0d892aa816d48560a90560571e) and D30344435 (https://github.com/facebook/pyre-check/commit/21689a0f47e08fbd561d50074eb954f2e58c63ee) cause the ocaml side to stop
spitting out annotation information for irrelevant modules, which causes
python to no longer produce `.pyi` files.

Even worse, because of the way that `paths_to_modify` was overloaded to act as
both a flag for in-place annotations and a list of modules (this behavior
predated recent infer work, but I remember thinking this was dangerous and
wanting to fix it but at the time there were many higher-priority things
to fix on the python side and then I forgot).

As a result, when I manually tested using `--print-only`
with no modules to annotate restrictions everything worked as expected,
ocaml got `None` as the modify paths and didn't filter. But in production
(e.g. when automatically running `pyre infer` before `pysa`), we wind up
passing an empty list to ocaml and then filtering out everything.

Reviewed By: arthaud

Differential Revision: D30697425

fbshipit-source-id: 5b5b8727a7a9637f096d129c1ed3401b4029bb41