Pin a node package version exactly (#469)
Summary:
Pull Request resolved: https://github.com/facebook/pyre-check/pull/469

It looks to me like there was an incorrect version
bump in one of the docusaurus modules - the latest
version of
`docusaurus-plugin-internaldocs-fb` marked as
compatible with 0.9.2 is in fact depending on breaking
changes in `docusaurus` core, which is breaking both
pyre's github CI jobs and local development after a fresh
`yarn install`.

The specific problem is that the floating version of
`internaldocs` is making use of
'docusaurus/useIsBrowser', which does not exist
in the docusaurus version range we are using. But the details
aren't very interesting, it's just a version mismatch problem.

As a quick fix, I'm pinning to 0.9.2 exactly. I'm also filing
a ticket to upgrade to the latest everything, which may take
some time; in my first attempt I ran into validation errors.

Reviewed By: shannonzhu

Differential Revision: D30711366

fbshipit-source-id: 7306f4b6844c462ba7ba664ac0aac0576aec0297