Avoid capture_output in scripts
Summary:
The `capture_output` argument of `subprocess.call` does not exist
in python3.6, which pyre-check is supposed to support. See
https://stackoverflow.com/questions/53209127/subprocess-unexpected-keyword-argument-capture-output/53209196
for some discussion and links to docs / code.

Inside the pyre client code, I intend to solve this by extracting
subprocess interactions into a utility module (which is something
I've wanted to do anyway to isolate the lint errors we get from
remote-code-execution warnings). But inside of scripts we're
using bare standard library python, so we have to solve the problem
by hand.

Reviewed By: dkgi

Differential Revision: D30737124

fbshipit-source-id: 977811b12d448267999a96a7211a553f304ade70