Fix illegal navbar item
Summary:
All entries in the `items` list have to be valid, so you can use the JS spread operator (`...`), to achieve this.

I believe this is equivalent to the splat (`*`) operator in python.

It's expanding a list of size one when internal, inserting a single element, and its expanding an empty list when external, effectively inserting no elements.

Context: https://fb.workplace.com/groups/docusaurus/posts/1425817471123703

Differential Revision: D30766078

fbshipit-source-id: 87796f839afc0ee1e828ec2437fb43c2d76c5b57