Hook github actions up to tox
Summary:
In a previous commit, we added a `[tool.tox]` section to
pyproject.toml so that we can run tests and lints on local
code in a tox-controlled way.

This commit uses that same setup to power the github tests,
following the pattern described in github actions docs at
https://docs.github.com/en/actions/guides/building-and-testing-python#running-tests-with-tox
to convert from the github python matrix to tox.

Differential Revision: D30788133

fbshipit-source-id: f7e83761d15814292920d077429d8d43a5c67834