Fix illegal navbar item (#470)
Summary:
Pull Request resolved: https://github.com/facebook/pyre-check/pull/470

All entries in the `items` list have to be valid, so you can use the JS spread operator (`...`), to achieve this.

I believe this is equivalent to the splat (`*`) operator in python.

It's expanding a list of size one when internal, inserting a single element, and its expanding an empty list when external, effectively inserting no elements.

Context: https://fb.workplace.com/groups/docusaurus/posts/1425817471123703

Reviewed By: stroxler

Differential Revision: D30766078

fbshipit-source-id: 647977e86673083c6ee4a3b94405652716052a85