Hook github actions up to tox (#471)
Summary:
Pull Request resolved: https://github.com/facebook/pyre-check/pull/471

In a previous commit, we added a `[tool.tox]` section to
pyproject.toml so that we can run tests and lints on local
code in a tox-controlled way.

This commit uses that same setup to power the github tests,
following the pattern described in github actions docs at
https://docs.github.com/en/actions/guides/building-and-testing-python#running-tests-with-tox
to convert from the github python matrix to tox.

Reviewed By: 0xedward

Differential Revision: D30788133

fbshipit-source-id: 08f4531d2d853a9043a49929cbe0984252a5e5f9