Add linting to tox
Summary:
It's customary to use tox not only for tests but also for linting;
this commit adds tox to our lint setup.

I went ahead and included `.tox` in the `.flake8` excludes because
open-source contributors are likely to run flake8 manually and we
shouldn't have a config that causes that to hang (which it does after
running tox because the .tox directory contains 4 virtual environments).

Reviewed By: dkgi

Differential Revision: D30778909

fbshipit-source-id: b529ef2122a35c4f7d1743fd106a7ccb996cf00c