Move ViaTypeOf and ViaValueOf features out of the breadcrumb set
Summary:
One of the most expensive operations in the analysis is to handle a call site: we need to get the model of the callee and propagate all sources, sinks and tito.
During this step, we currently iterate on all breadcrumbs of the callee model in order to find `ViaValueOf` and `ViaTypeOf` features, and instantiate those `via-type:..` breadcrumbs.
This can be quite costly since the breadcrumb set can get quite large.

To prevent this, this patch moves the `ViaTypeOf` and `ViaValueOf` features into a different set in `FlowDetails`, so that we only iterate on those, and not the whole set of breadcrumbs. This might bring performance benefits. It also makes more sense to use a normal set instead of an over-under set for these, since we do not care whether `ViaXXXOf` is present on all flows or not.

Reviewed By: dkgi

Differential Revision: D30761555

fbshipit-source-id: 99a56072b8ad65833cc23905267d31338e181625