Add functions to add or get features out of taint representations
Summary:
A very recurrent pattern in our codebase is to gather or add features or breadcrumb on a taint abstraction. For now, this is usually done with a fold on the abstract domain.
Let's add helper functions to add or gather breadcrumbs and features.
This will make refactoring easier when changing the abstraction for the feature set, for instance splitting via-features and breadcrumbs, or implementing local features.

Reviewed By: dkgi

Differential Revision: D30761500

fbshipit-source-id: 5e7dc8e180b38b302b470cdb32ca40b9ffbfdb2a