Rename leaf into kind
Summary:
The term `leaf` is pretty ambiguous in our codebase. It is used in very various ways:
* To denote the source or sink kind (e.g, `UserControlled`). It is also called `leaf kind` sometimes.
* To denote a leaf name, i.e the callable that originated a source or sink, i.e the first frame or last frame of the trace.
* To denote a leaf taint, i.e the first frame or last frame of the trace.
* To denote a leaf in a binary tree.

To prevent confusion, I propose to rename the first use into `kind`. This is a lot more explicit in my opinion, and it is also the term we use in our user interface.

Reviewed By: dkgi

Differential Revision: D30758453

fbshipit-source-id: a99bf4c7883e19c35bd35e800d19b2e5e8b8ecd0